% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{createSmokingSettings}
\alias{createSmokingSettings}
\title{createSmokingSettings}
\usage{
createSmokingSettings(startDay = -365, endDay = 0, analysisId = 639)
}
\arguments{
\item{startDay}{The days prior to target index to look for a patient being in the smoking cohort}

\item{endDay}{The days after to target index to look for a patient being in the smoking cohort}

\item{analysisId}{A covariate analysis ID to use for the smoking feature}
}
\value{
A covariate setting that will create a smoking covariate based on whether a patient
has observation concept id 40766362 and the value (current, previous, not smoker)
}
\description{
Create a covariate setting for the smoking variable (warning: this uses a concept ID that 
 could change as the vocab changes over time)
}
\details{
The function creates a list of cohort subset operators based on the specified inputs
}
\examples{
smokingCovSet <- createSmokingSettings(
   startDay = -365,
   endDay = 0,
   analysisId = 639
   )
}
