% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developFunction.R
\name{developLungCancerEHRmodel}
\alias{developLungCancerEHRmodel}
\title{Develop the model}
\usage{
developLungCancerEHRmodel(
  databaseDetails,
  outputFolder,
  createCohorts = F,
  runAnalyses = F,
  sampleSize = NULL,
  logSettings
)
}
\arguments{
\item{databaseDetails}{Database details created using \code{PatientLevelPrediction::createDatabaseDetails()}}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{createCohorts}{Create the cohortTable table with the target population and outcome cohorts?}

\item{runAnalyses}{Run the model development}

\item{sampleSize}{The number of patients in the target cohort to sample (if NULL uses all patients)}

\item{logSettings}{The log setting \code{PatientLevelPrediction::createLogSettings()}}
}
\description{
Develop the model
}
\details{
This function executes the LungCancerPrognsotic model development.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")
                                             
 databaseDetails <- PatientLevelPrediction::createDatabaseDetails(
 connectionDetails = connectionDetails,
 cdmDatabaseSchema = cdmDatabaseSchema,
 cdmDatabaseName = cdmDatabaseName,
 tempEmulationSchema = tempEmulationSchema,
 cohortDatabaseSchema = cohortDatabaseSchema,
 cohortTable = cohortTable,
 outcomeDatabaseSchema = cohortDatabaseSchema,
 outcomeTable = cohortTable,
 cdmVersion = cdmVersion
 )  
 
 logSettings <- PatientLevelPrediction::createLogSettings(
 verbosity = "INFO",
 timeStamp = T,
 logName = 'skeletonPlp'
 )                                          

developLungCancerEHRmodel(
databaseDetails = databaseDetails,
        outputFolder = "c:/temp/study_results", 
        createCohorts = T,
        runAnalyses = T,
        logSettings = logSettings
        )
}

}
