% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{getSmokingCovariateData}
\alias{getSmokingCovariateData}
\title{Smoking variable}
\usage{
getSmokingCovariateData(
  connection,
  oracleTempSchema = NULL,
  cdmDatabaseSchema,
  cohortTable = "#cohort_person",
  cohortId = -1,
  cdmVersion = "5",
  rowIdField = "subject_id",
  covariateSettings,
  aggregated = FALSE
)
}
\arguments{
\item{oracleTempSchema}{bla bla}

\item{cdmDatabaseSchema}{bla bla}

\item{cohortTable}{bla bla}

\item{cohortId}{bla bla}

\item{cdmVersion}{bla bla}

\item{rowIdField}{bla bla}

\item{covariateSettings}{bla bla}

\item{aggregated}{bla bla}
}
\description{
Smoking variable
}
\details{
This will apply the model to new data and estimate the performance
}
