% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LungCancerPrognostic.R
\docType{package}
\name{LungCancerPrognostic}
\alias{LungCancerPrognostic-package}
\alias{LungCancerPrognostic}
\title{LungCancerPrognostic}
\description{
A package for running prediction model development studies using data in the OMOP CDM
}
