% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateFunction.R
\name{validateLungCancerEHRmodel}
\alias{validateLungCancerEHRmodel}
\title{Apply the model to new OMOP CDM data and get external performance}
\usage{
validateLungCancerEHRmodel(
  databaseDetails,
  validationRestrictPlpDataSettings,
  settings = PatientLevelPrediction::createValidationSettings(recalibrate =
    "weakRecalibration"),
  outputFolder
)
}
\arguments{
\item{databaseDetails}{The databaseDetails}

\item{validationRestrictPlpDataSettings}{Whether to sample from the target population}

\item{settings}{Whether to recalibrate}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}
}
\description{
Apply the model to new OMOP CDM data and get external performance
}
\details{
This will apply the model to new data and estimate the performance
}
